requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.14.6
chardet==3.0.4
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.4
cryptography==3.4.8
decorator==4.1.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-configurations==2.2
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-registration==3.2
djangorestframework==3.12.4
drf-yasg==1.21.7
idna==2.6
inflection==0.5.1
ipython==5.5.0
ipython_genutils==0.2.0
keyring==10.6.0
keyrings.alt==3.0
netifaces==0.10.4
oauthlib==3.1.1
packaging==21.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
pycparser==2.20
pycrypto==2.6.1
Pygments==2.2.0
PyGObject==3.26.1
PyJWT==2.1.0
pyparsing==3.1.1
python-apt==1.6.5+ubuntu0.3
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.25
PyYAML==6.0.1
requests==2.18.4
requests-oauthlib==1.3.0
SecretStorage==2.3.1
simplegeneric==0.8.1
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.7
traitlets==4.3.2
typing-extensions==3.10.0.1
uritemplate==4.1.1
urllib3==1.22
wcwidth==0.1.7
